remove PlanetScaleMysqlDatabase type, collapse functionality into PlanetScaleEdgeDatabase
